+++ tests/test_pycrud.py
@@ -2,4 +2,4 @@ from pycurd import __version__
 
 
 def test_version():
-    assert __version__ == '0.1.13'
+    assert __version__ == '0.1.14'